+++ src/proxy/h2_client.rs
@@ -1,0 +1,268 @@
+use crate::config;
+use crate::proxy::Error;
+use bytes::Buf;
+use bytes::Bytes;
+use futures_core::ready;
+use h2::client::{Connection, SendRequest};
+use h2::{Reason, SendStream};
+use http::Request;
+use std::io::Cursor;
+use std::pin::Pin;
+use std::sync::atomic::{AtomicU16, Ordering};
+use std::sync::Arc;
+use std::task::{Context, Poll};
+use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
+use tokio::net::TcpStream;
+use tokio::sync::watch::Receiver;
+use tokio_rustls::client::TlsStream;
+use tracing::{debug, error, trace};
+
+// H2Stream represents an active HTTP2 stream. Consumers can only Read/Write
+pub struct H2Stream {
+    send_stream: h2::SendStream<SendBuf>,
+    recv_stream: h2::RecvStream,
+    buf: Bytes,
+    active_count: Arc<AtomicU16>,
+}
+
+impl H2Stream {
+    fn write_slice(&mut self, buf: &[u8], end_of_stream: bool) -> Result<(), std::io::Error> {
+        let send_buf: SendBuf = Cursor::new(buf.into());
+        self.send_stream
+            .send_data(send_buf, end_of_stream)
+            .map_err(h2_to_io_error)
+    }
+}
+
+impl Drop for H2Stream {
+    fn drop(&mut self) {
+        let left = self.active_count.fetch_sub(1, Ordering::SeqCst);
+        trace!("dropping h2stream, has {} active streams left", left - 1);
+    }
+}
+
+impl AsyncRead for H2Stream {
+    fn poll_read(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        read_buf: &mut ReadBuf<'_>,
+    ) -> Poll<std::io::Result<()>> {
+        if self.buf.is_empty() {
+            self.buf = loop {
+                match ready!(self.recv_stream.poll_data(cx)) {
+                    None => return Poll::Ready(Ok(())),
+                    Some(Ok(buf)) if buf.is_empty() && !self.recv_stream.is_end_stream() => {
+                        continue
+                    }
+                    Some(Ok(buf)) => {
+                        // TODO: implement ping
+                        // self.ping.record_data(buf.len());
+                        break buf;
+                    }
+                    Some(Err(e)) => {
+                        return Poll::Ready(match e.reason() {
+                            Some(Reason::NO_ERROR) | Some(Reason::CANCEL) => Ok(()),
+                            Some(Reason::STREAM_CLOSED) => {
+                                Err(std::io::Error::new(std::io::ErrorKind::BrokenPipe, e))
+                            }
+                            _ => Err(h2_to_io_error(e)),
+                        })
+                    }
+                }
+            };
+        }
+        let cnt = std::cmp::min(self.buf.len(), read_buf.remaining());
+        read_buf.put_slice(&self.buf[..cnt]);
+        self.buf.advance(cnt);
+        let _ = self.recv_stream.flow_control().release_capacity(cnt);
+        Poll::Ready(Ok(()))
+    }
+}
+
+impl AsyncWrite for H2Stream {
+    fn poll_write(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &[u8],
+    ) -> Poll<Result<usize, std::io::Error>> {
+        if buf.is_empty() {
+            return Poll::Ready(Ok(0));
+        }
+        self.send_stream.reserve_capacity(buf.len());
+
+        // We ignore all errors returned by `poll_capacity` and `write`, as we
+        // will get the correct from `poll_reset` anyway.
+        let cnt = match ready!(self.send_stream.poll_capacity(cx)) {
+            None => Some(0),
+            Some(Ok(cnt)) => self.write_slice(&buf[..cnt], false).ok().map(|()| cnt),
+            Some(Err(_)) => None,
+        };
+
+        if let Some(cnt) = cnt {
+            return Poll::Ready(Ok(cnt));
+        }
+
+        Poll::Ready(Err(h2_to_io_error(
+            match ready!(self.send_stream.poll_reset(cx)) {
+                Ok(Reason::NO_ERROR) | Ok(Reason::CANCEL) | Ok(Reason::STREAM_CLOSED) => {
+                    return Poll::Ready(Err(std::io::ErrorKind::BrokenPipe.into()))
+                }
+                Ok(reason) => reason.into(),
+                Err(e) => e,
+            },
+        )))
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), std::io::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn poll_shutdown(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Result<(), std::io::Error>> {
+        if self.write_slice(&[], true).is_ok() {
+            return Poll::Ready(Ok(()));
+        }
+
+        Poll::Ready(Err(h2_to_io_error(
+            match ready!(self.send_stream.poll_reset(cx)) {
+                Ok(Reason::NO_ERROR) => return Poll::Ready(Ok(())),
+                Ok(Reason::CANCEL) | Ok(Reason::STREAM_CLOSED) => {
+                    return Poll::Ready(Err(std::io::ErrorKind::BrokenPipe.into()))
+                }
+                Ok(reason) => reason.into(),
+                Err(e) => e,
+            },
+        )))
+    }
+}
+
+pub type SendBuf = Cursor<Box<[u8]>>;
+
+fn h2_to_io_error(e: h2::Error) -> std::io::Error {
+    if e.is_io() {
+        e.into_io().unwrap()
+    } else {
+        std::io::Error::new(std::io::ErrorKind::Other, e)
+    }
+}
+
+pub async fn spawn_connection(
+    cfg: Arc<config::Config>,
+    s: TlsStream<TcpStream>,
+    driver_drain: Receiver<bool>,
+) -> Result<H2ConnectClient, Error> {
+    let mut builder = h2::client::Builder::new();
+    builder
+        .initial_window_size(cfg.window_size)
+        .initial_connection_window_size(cfg.connection_window_size)
+        .max_frame_size(cfg.frame_size)
+        .initial_max_send_streams(cfg.pool_max_streams_per_conn as usize)
+        .max_header_list_size(1024 * 16)
+        .max_send_buffer_size(1024 * 1024)
+        .enable_push(false);
+
+    let (send_req, connection) = builder
+        .handshake::<_, SendBuf>(s)
+        .await
+        .map_err(Error::Http2Handshake)?;
+
+    // We store max as u16, so if they report above that max size we just cap at u16::MAX
+    let max_allowed_streams = std::cmp::min(
+        cfg.pool_max_streams_per_conn,
+        connection
+            .max_concurrent_send_streams()
+            .try_into()
+            .unwrap_or(u16::MAX),
+    );
+    // spawn a task to poll the connection and drive the HTTP state
+    // if we got a drain for that connection, respect it in a race
+    // it is important to have a drain here, or this connection will never terminate
+    tokio::spawn(async move {
+        drive_connection(connection, driver_drain).await;
+    });
+
+    let c = H2ConnectClient {
+        sender: send_req,
+        stream_count: Arc::new(AtomicU16::new(0)),
+        max_allowed_streams,
+    };
+    Ok(c)
+}
+
+#[derive(Debug, Clone)]
+// H2ConnectClient is a wrapper abstracting h2
+pub struct H2ConnectClient {
+    sender: SendRequest<SendBuf>,
+    pub max_allowed_streams: u16,
+    stream_count: Arc<AtomicU16>,
+}
+
+impl H2ConnectClient {
+    pub fn at_max_streamcount(&self) -> bool {
+        let curr_count = self.stream_count.load(Ordering::Relaxed);
+        trace!(
+            "checking streamcount: {curr_count} >= {}",
+            self.max_allowed_streams
+        );
+        curr_count >= self.max_allowed_streams
+    }
+
+    pub async fn send_request(&mut self, req: http::Request<()>) -> Result<H2Stream, Error> {
+        let cur = self.stream_count.fetch_add(1, Ordering::SeqCst);
+        trace!(current_streams = cur, "sending request");
+        let (send, recv) = match self.internal_send(req).await {
+            Ok(r) => r,
+            Err(e) => {
+                // Request failed, so drop the stream now
+                self.stream_count.fetch_sub(1, Ordering::SeqCst);
+                return Err(e);
+            }
+        };
+
+        let h2 = H2Stream {
+            send_stream: send,
+            recv_stream: recv,
+            buf: Bytes::new(),
+            active_count: self.stream_count.clone(),
+        };
+        Ok(h2)
+    }
+
+    // helper to allow us to handle errors once
+    async fn internal_send(
+        &mut self,
+        req: Request<()>,
+    ) -> Result<(SendStream<SendBuf>, h2::RecvStream), Error> {
+        let (response, stream) = self.sender.send_request(req, false)?;
+        let response = response.await?;
+        if response.status() != 200 {
+            return Err(Error::HttpStatus(response.status()));
+        }
+        Ok((stream, response.into_body()))
+    }
+}
+
+async fn drive_connection<S, B>(c: Connection<S, B>, mut driver_drain: Receiver<bool>)
+where
+    S: AsyncRead + AsyncWrite + Send + Unpin,
+    B: Buf,
+{
+    // TODO: ping pong
+    tokio::select! {
+        _ = driver_drain.changed() => {
+            debug!("draining outer HBONE connection");
+        }
+        res = c => {
+            match res {
+                Err(e) => {
+                    error!("Error in HBONE connection handshake: {:?}", e);
+                }
+                Ok(_) => {
+                    debug!("done with HBONE connection handshake: {:?}", res);
+                }
+            }
+        }
+    }
+}
